+++ Box.V2.Test/BoxRetentionPoliciesManagerTest.cs
@@ -192,5 +192,68 @@ namespace Box.V2.Test
             Assert.AreEqual("Contract.pdf", results.Entries[0].File.Name);
             Assert.AreEqual("12345", results.Entries[0].WinningRetentionPolicy.Id);
         }
+
+        [TestMethod]
+        public async Task GetFilesUnderRetentionForAssignment_ValidResponse()
+        {
+            /*** Arrange ***/
+            string retentionPolicyAssignmentId = "12345";
+            string responseString = "{ \"entries\": [{ \"id\": 12345, \"etag\": 1, \"type\": \"file\", \"sequence_id\": 3, \"name\": \"Contract.pdf\", \"sha1\": \"85136C79CBF9FE36BB9D05D0639C70C265C18D37\", \"file_version\": { \"id\": 123456, \"type\": \"file_version\", \"sha1\": \"134b65991ed521fcfe4724b7d814ab8ded5185dc\" }, \"applied_at\": \"2012-12-12T10:53:43-08:00\" } ], \"limit\": 1000, \"marker\": \"some marker\" }";
+            IBoxRequest boxRequest = null;
+            Handler.Setup(h => h.ExecuteAsync<BoxCollectionMarkerBased<BoxFile>>(It.IsAny<IBoxRequest>()))
+                .Returns(Task.FromResult<IBoxResponse<BoxCollectionMarkerBased<BoxFile>>>(new BoxResponse<BoxCollectionMarkerBased<BoxFile>>()
+                {
+                    Status = ResponseStatus.Success,
+                    ContentString = responseString
+                }))
+                .Callback<IBoxRequest>(r => boxRequest = r);
+
+            /*** Act ***/
+            BoxCollectionMarkerBased<BoxFile> results = await _retentionPoliciesManager.GetFilesUnderRetentionForAssignmentAsync(retentionPolicyAssignmentId);
+
+            /*** Assert ***/
+
+            // Request check
+            Assert.IsNotNull(boxRequest);
+            Assert.AreEqual(RequestMethod.Get, boxRequest.Method);
+            Assert.AreEqual(retentionPolicyAssignmentId, boxRequest.Parameters["retention_policy_assignment_id"]);
+
+            // Response check
+            Assert.AreEqual("12345", results.Entries[0].Id);
+            Assert.AreEqual("Contract.pdf", results.Entries[0].Name);
+            Assert.AreEqual("file", results.Entries[0].Type);
+        }
+
+        [TestMethod]
+        public async Task GetFileVersionsUnderRetentionForAssignment_ValidResponse()
+        {
+            /*** Arrange ***/
+            string retentionPolicyAssignmentId = "12345";
+            string responseString = "{ \"entries\": [{ \"id\": 12345, \"etag\": 1, \"type\": \"file_version\", \"sequence_id\": 3, \"name\": \"Contract.pdf\", \"sha1\": \"85136C79CBF9FE36BB9D05D0639C70C265C18D37\", \"file_version\": { \"id\": 123456, \"type\": \"file_version\", \"sha1\": \"134b65991ed521fcfe4724b7d814ab8ded5185dc\" }, \"applied_at\": \"2012-12-12T10:53:43-08:00\" } ], \"limit\": 1000, \"marker\": \"some marker\" }";
+            IBoxRequest boxRequest = null;
+            Handler.Setup(h => h.ExecuteAsync<BoxCollectionMarkerBased<BoxFileVersion>>(It.IsAny<IBoxRequest>()))
+                .Returns(Task.FromResult<IBoxResponse<BoxCollectionMarkerBased<BoxFileVersion>>>(new BoxResponse<BoxCollectionMarkerBased<BoxFileVersion>>()
+                {
+                    Status = ResponseStatus.Success,
+                    ContentString = responseString
+                }))
+                .Callback<IBoxRequest>(r => boxRequest = r);
+
+            /*** Act ***/
+            BoxCollectionMarkerBased<BoxFileVersion> results = await _retentionPoliciesManager.GetFileVersionsUnderRetentionForAssignmentAsync(retentionPolicyAssignmentId);
+
+            /*** Assert ***/
+
+            // Request check
+            Assert.IsNotNull(boxRequest);
+            Assert.AreEqual(RequestMethod.Get, boxRequest.Method);
+            Assert.AreEqual(retentionPolicyAssignmentId, boxRequest.Parameters["retention_policy_assignment_id"]);
+
+            // Response check
+            Assert.AreEqual("12345", results.Entries[0].Id);
+            Assert.AreEqual("Contract.pdf", results.Entries[0].Name);
+            Assert.AreEqual("file_version", results.Entries[0].Type);
+            Assert.AreEqual("file_version", results.Entries[0].FileVersion.Type);
+        }
     }
 }