+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next release
 - Extract interfaces for BoxClient and Managers to improve testability ([#603](https://github.com/box/box-windows-sdk-v2/pull/603))
+- Add ability to get files under retention for assignment and file versions under retention for assignment ([#734](https://github.com/box/box-windows-sdk-v2/pull/734))
 
 ## 3.26.0 [2021-04-01]